[CHORE] 사용자 학습시간 응답객체 처리 방식 변경
