+++ smeem-api/src/main/java/com/smeem/api/member/service/dto/response/MemberGetServiceResponse.java
@@ -6,8 +6,15 @@ import com.smeem.domain.goal.model.GoalType;
 import com.smeem.domain.member.model.LangType;
 import com.smeem.domain.member.model.Member;
 import com.smeem.domain.plan.model.Plan;
+import com.smeem.domain.training.exception.TrainingTimeException;
+import com.smeem.domain.training.model.TrainingTime;
 import lombok.Builder;
+import lombok.val;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.smeem.common.code.failure.TrainingTimeFailureCode.NOT_SET_TRAINING_TIME;
 import static lombok.AccessLevel.PRIVATE;
 
 
@@ -28,19 +35,43 @@ public record MemberGetServiceResponse(
             GoalGetServiceResponse goal,
             Member member,
             Plan trainingPlan,
-            TrainingTimeServiceResponse trainingTime,
             BadgeServiceResponse badge
-    )  {
+    ) {
+
         return MemberGetServiceResponse.builder()
                 .username(member.getUsername())
-                .goalType(goal.goalType())
+                .goalType(member.getGoal())
                 .way(goal.way())
                 .detail(goal.detail())
                 .targetLangType(member.getTargetLang())
                 .hasPushAlarm(member.isHasPushAlarm())
-                .trainingTime(trainingTime)
+                .trainingTime(generateTrainingTimeResponse(member.getTrainingTimes()))
                 .badge(badge)
                 .trainingPlan(PlanServiceResponse.of(trainingPlan))
                 .build();
     }
+
+    private static String getDays(List<TrainingTime> trainingTimes) {
+        return trainingTimes.stream()
+                .map(trainingTime -> trainingTime.getDay().name())
+                .distinct()
+                .collect(Collectors.joining(","));
+    }
+
+    private static TrainingTimeServiceResponse generateTrainingTimeResponse(List<TrainingTime> trainingTimes) {
+        if (trainingTimes.isEmpty()) {
+            return TrainingTimeServiceResponse.of("", 22, 0);
+        }
+        val trainingTime = getOneTrainingTime(trainingTimes);
+        return TrainingTimeServiceResponse.of(
+                getDays(trainingTimes),
+                trainingTime.getHour(),
+                trainingTime.getMinute());
+    }
+
+    private static TrainingTime getOneTrainingTime(List<TrainingTime> trainingTimes) {
+        return trainingTimes.stream().findFirst().orElseThrow(
+                () -> new TrainingTimeException(NOT_SET_TRAINING_TIME));
+    }
+
 }
